Automated form submission notification not sending
Automated form submission notification not sending after updating to the most recent Mura version.

The proposed change is what we are implementing to get by until a finalized fix can be put in place. I just changed line 262 to use the getResponseSendTo method, rather than getSendTo so that the proper email address data is passed through to the mailer submit method.